Decompose banking economics boundary
